# rexapi/2_1/swagger.yaml
consumes:
  - application/json
definitions:
  AuthenticationRequest:
    description: ''
    properties:
      Secret:
        $ref: '#/definitions/UserAuthenticationInfo'
      User:
        $ref: '#/definitions/User'
    required:
      - User
      - Secret
    type: object
  AuthenticationToken:
    description: >-
      The spec permits you to use any token format you like. You could, for
      example, look into JSON Web Tokens ("JWT", pronounced "jots"):
      https://jwt.io.
    type: string
...
  Package:
    properties:
      data:
        $ref: '#/definitions/PackageData'
      metadata:
        $ref: '#/definitions/PackageMetadata'
    required:
      - metadata
      - data
    type: object
  PackageData:
    description: |-
      This is a "union" type.
      - On package upload, either Content or URL should be set.
      - On package update, exactly one field should be set.
      - On download, the Content field should be set.
    properties:
      Content:
        description: >-
          Package contents. This is the zip file uploaded by the user. (Encoded
          as text using a Base64 encoding).


          This will be a zipped version of an npm package's GitHub repository,
          minus the ".git/" directory." It will, for example, include the
          "package.json" file that can be used to retrieve the project homepage.


          See
          https://docs.npmjs.com/cli/v7/configuring-npm/package-json#homepage.
        type: string
      JSProgram:
        description: A JavaScript program (for use with sensitive modules).
        type: string
      URL:
        description: Package URL (for use in public ingest).
        type: string
    type: object
...
host: package-registry-461.appspot.com
info:
  contact:
    email: davisjam@purdue.edu
    name: Prof. Davis
    url: 'http://davisjam.github.io'
  description: 'API for ECE 461/Spring 2023/Project 2: A Trustworthy Module Registry'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
  termsOfService: 'http://swagger.io/terms/'
  title: ECE 461 - Spring 2023 - Project 2
  version: 2.0.0
paths:
  "/authenticate":
    put:
      description: Create an access token.
      operationId: CreateAuthToken
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        200:
          description: Return an AuthenticationToken.
          schema:
            $ref: '#/definitions/AuthenticationToken'
        400:
          description: >-
            There is missing field(s) in the AuthenticationRequest or it is
            formed improperly.
        401:
          description: The user or password is invalid.
        501:
          description: This system does not support authentication.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/AuthenticationRequest'
  "/package":
    parameters:
      - description: >-
          The spec permits you to use any token format you like. You could, for
          example, look into JSON Web Tokens ("JWT", pronounced "jots"):
          https://jwt.io.
        in: header
        name: X-Authorization
        required: true
        type: string
    post:
      consumes:
        - application/json
      operationId: PackageCreate
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/PackageData'
      produces:
        - application/json
      responses:
        201:
          description: Success. Check the ID in the returned metadata for the official ID.
          schema:
            $ref: '#/definitions/Package'
        400:
          description: >-
            There is missing field(s) in the PackageData/AuthenticationToken or
            it is formed improperly, or the AuthenticationToken is invalid.
        409:
          description: Package exists already.
        424:
          description: Package is not uploaded due to the disqualified rating.
  "/package/byName/{name}":
    delete:
      operationId: PackageByNameDelete
      parameters:
        - description: |-
            Name of a package.

            - Names should only use typical "keyboard" characters.
            - The name "*" is reserved. See the `/packages` API for its meaning.
          in: path
          name: name
          required: true
          type: string
      responses:
        200:
          description: Package is deleted.
        400:
          description: >-
            There is missing field(s) in the PackageName/AuthenticationToken or
            it is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: Package does not exist.
      summary: Delete all versions of this package.
    get:
      description: Return the history of this package (all versions).
      operationId: PackageByNameGet
      parameters:
        - description: >-
            The spec permits you to use any token format you like. You could,
            for example, look into JSON Web Tokens ("JWT", pronounced "jots"):
            https://jwt.io.
          in: header
          name: X-Authorization
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: Return the package history.
          schema:
            items:
              $ref: '#/definitions/PackageHistoryEntry'
            type: array
        400:
          description: >-
            There is missing field(s) in the PackageName/AuthenticationToken or
            it is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: No such package.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
    parameters:
      - description: |-
          Name of a package.

          - Names should only use typical "keyboard" characters.
          - The name "*" is reserved. See the `/packages` API for its meaning.
        in: path
        name: name
        required: true
        type: string
      - description: >-
          The spec permits you to use any token format you like. You could, for
          example, look into JSON Web Tokens ("JWT", pronounced "jots"):
          https://jwt.io.
        in: header
        name: X-Authorization
        required: true
        type: string
  "/package/byRegEx":
    parameters:
      - description: >-
          The spec permits you to use any token format you like. You could, for
          example, look into JSON Web Tokens ("JWT", pronounced "jots"):
          https://jwt.io.
        in: header
        name: X-Authorization
        required: true
        type: string
    post:
      consumes:
        - text/plain
      description: >-
        Search for a package using regular expression over package names and
        READMEs. This is similar to search by name.
      operationId: PackageByRegExGet
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/PackageRegEx'
      produces:
        - application/json
      responses:
        200:
          description: Return a list of packages.
          schema:
            items:
              $ref: '#/definitions/PackageMetadata'
            type: array
        400:
          description: >-
            There is missing field(s) in the PackageRegEx/AuthenticationToken or
            it is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: No package found under this regex.
      summary: Get any packages fitting the regular expression.
  "/package/{id}":
    delete:
      operationId: PackageDelete
      parameters:
        - description: Package ID
          in: path
          name: id
          required: true
          type: string
      responses:
        200:
          description: Package is deleted.
        400:
          description: >-
            There is missing field(s) in the PackageID/AuthenticationToken or it
            is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: Package does not exist.
      summary: Delete this version of the package.
    get:
      description: Return this package.
      operationId: PackageRetrieve
      parameters:
        - description: ID of package to fetch
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        200:
          description: Return the package.
          schema:
            $ref: '#/definitions/Package'
        400:
          description: >-
            There is missing field(s) in the PackageID/AuthenticationToken or it
            is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: Package does not exist.
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
      summary: Interact with the package with this ID
    parameters:
      - description: Package ID
        in: path
        name: id
        required: true
        type: string
      - description: >-
          The spec permits you to use any token format you like. You could, for
          example, look into JSON Web Tokens ("JWT", pronounced "jots"):
          https://jwt.io.
        in: header
        name: X-Authorization
        required: true
        type: string
    put:
      consumes:
        - application/json
      description: >-
        The name, version, and ID must match.


        The package contents (from PackageData) will replace the previous
        contents.
      operationId: PackageUpdate
      parameters:
        - description: ''
          in: path
          name: id
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: '#/definitions/Package'
            
      responses:
        200:
          description: Version is updated.
        400:
          description: >-
            There is missing field(s) in the PackageID/AuthenticationToken or it
            is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: Package does not exist.
      summary: Update this content of the package.
  "/package/{id}/rate":
    get:
      operationId: PackageRate
      parameters: []
      produces:
        - application/json
      responses:
        200:
          description: >-
            Return the rating. Only use this if each metric was computed
            successfully.
          schema:
            $ref: '#/definitions/PackageRating'
        400:
          description: >-
            There is missing field(s) in the PackageID/AuthenticationToken or it
            is formed improperly, or the AuthenticationToken is invalid.
        404:
          description: Package does not exist.
        500:
          description: The package rating system choked on at least one of the metrics.
    parameters:
      - in: path
        name: id
        required: true
        type: string
      - description: >-
          The spec permits you to use any token format you like. You could, for
          example, look into JSON Web Tokens ("JWT", pronounced "jots"):
          https://jwt.io.
        in: header
        name: X-Authorization
        required: true
        type: string
...
produces:
  - application/json
schemes:
  - http
securityDefinitions:
  UserAndPasswordYieldsToken:
    description: Use the /authentication endpoint to obtain an authorization token.
    in: header
    name: X-Authorization
    type: apiKey
swagger: '2.0'
...